New-style Ingredient Refactor
Ingredients are now based on Godot resources and rather than sharing a spritesheet, have their own sprites (added in an earlier commit). Additionally, they have Name, Quality, and State properties, with the plan to change textures based on changing the state, but this hasn't been tested yet.

The old-style ingredient classes and scenes have been removed, except for IngredientButton, which has been updated to work with the new-style ingredients.